requirements/base.txt
-e git+https://github.com/randlet/django-genericdropdown.git@473ff52610af659f7d2a3616a6e3322e21673b4d#egg=django_genericdropdown-master
-e git+https://github.com/randlet/django-listable.git@django_21#egg=django_listable
-e git+https://github.com/qatrackplus/django-q.git@qatrackplus#egg=django-q
-e git+https://github.com/randlet/django-recurrence.git@035ef961483a9ed58ecea6cac5dc8909b1227ef7#egg=django-recurrence
-e git+https://github.com/randlet/django-sql-explorer.git@196e1f5af2abed61f504878b89a0719ae0901ed8#egg=django-sql-explorer
Django==2.1.11
PyVirtualDisplay==0.2
beautifulsoup4==4.6.0
coreapi==2.3.3
coverage==4.5.1
django-admin-views==0.8.0
django-braces==1.13.0
django-contrib-comments==1.8.0
django-coverage==1.2.4
django-crispy-forms==1.6.0
django-debug-toolbar==2.0
django-dynamic-raw-id==2.5
django-extensions==2.1.4
django-filter==2.1.0
django-form-utils==1.0.3
django-formtools==2.1
django-registration==2.5.2
djangorestframework==3.9.1
djangorestframework-filters==1.0.0_dev0
django-widget-tweaks==1.4.1
freezegun==0.3.9
html5lib==1.0.1
markdown==2.6.11
matplotlib==2.2.2
numpy<=1.15.2
pandas>=0.22.0,<0.25
pep8==1.7.0
pydicom==1.1.0
scikit-image<0.14.0; python_version == '3.4'
scikit-learn<1.0; python_version == '3.4'
pylinac==2.1.0.1; python_version < '3.6'
-e git+https://github.com/qatrackplus/pylinac-qatrackplus.git@master#egg=pylinac ; python_version >= '3.6'
Pillow<6
pynliner==0.8.0
pytest-cov==2.6.1
pytest-django==3.4.5
pytest-sugar==0.9.1
pytest==4.1.1
python-dateutil==2.6.1
pytz==2018.4
reportlab==3.5.8
requests==2.21.0
scipy<=1.1.0
selenium==3.141.0
XlsxWriter==1.1.2
